Split Ruby functions from suppressions
This commit adds support for suppression files inside of the
ruby_memcheck codebase. This splits the list of Ruby functions from the
list of suppressions. This has several benefits:

1. The list of skipped Ruby functions now functions as a list of only
   Ruby functions that call back into Ruby (rather than also including
   false positives).
2. Suppression files will let us suppress non-memory leaks (e.g. this
   commit adds a Memcheck:Overlap suppression).
3. Suppression files will let us take advantage of the existing
   code for suppressing false positives in specific Ruby versions.

This commit introduces a deprecation in naming suppression files in the
format your_binary_name_ruby-3.supp to a simpler ruby-3.supp. The old
format remains supported but will output a warning.