Use only RUBY_FREE_AT_EXIT on supported Rubies
On Ruby 3.4 and later, we can exclusively use RUBY_FREE_AT_EXIT and
disable most of the heuristics. This removes some of the limitations of
the heuristics and makes ruby_memcheck more accurate.

This feature could cause more false-positives, so there is an escape
hatch by passing `use_only_ruby_free_at_exit: false` to the configuration.