Lower the limit on file descriptors with `ulimit -n`
Valgrind 3.21.0 will exit if the fd limit is too high. This change is
described here:

> https://bugs.kde.org/show_bug.cgi?id=465435

A quote from a maintainer in that issue is: "We're not going to change
Valgrind to handle insane rlimits that Docker presents", however it
looks like Ubuntu 24.10 sets this limit very high:

    $ uname -a
    Linux bepi 6.11.0-1004-lowlatency #4-Ubuntu SMP PREEMPT_DYNAMIC Mon
    Sep 30 10:54:09 UTC 2024 x86_64 x86_64 x86_64 GNU/Linux
    $ ulimit -n
    1073741816

I suspect this patch will not work on some systems for a variety of
reasons, but I wanted to start a conversation about how to address
this issue.